Use bounds.center in QuadTreeSpaceSplitter.IsPartOf check
In my testing the objects that were intersecting the borders of cells weren't
put in any cell, and there could be a lot of individual leftover objects that
don't become a part of any quadtree cell on lower levels when cells get small
enough. The intent of this commit's change is to have all objects become a part
of some cell, it doesn't matter if they're a bit out of it, as long as their
center is in the cell.