Changed UnityMeshSimplifier dependency from submodule to git repo
Unity's Package Manager doesn't work with submodules so you couldn't use this
fork's github URL as a package. Now it'll work if you also add the following
line to your project's manifest.json:

    "com.whinarn.unitymeshsimplifier": "https://github.com/Unity-Technologies/UnityMeshSimplifier.git#v2.3.4",

This commit was originally on the "backport to 2019.4" branch but it's probably
relevant for newer Unity versions as well.